fix: DB에 저장된 경로가 없을 때 조건 추가(#116)

+++ src/main/java/PodoeMarket/podoemarket/service/MypageService.java
@@ -179,10 +179,12 @@ public class MypageService {
         metadata.setContentType(files[0].getContentType());
 
         // 기존 파일 삭제
-        String formalS3Key = productRepo.findById(id).getImagePath();
+        if(productRepo.findById(id).getImagePath() != null){
+            final String formalS3Key = productRepo.findById(id).getImagePath();
 
-        if(amazonS3.doesObjectExist(bucket, formalS3Key)) {
-            amazonS3.deleteObject(bucket, formalS3Key);
+            if(amazonS3.doesObjectExist(bucket, formalS3Key)) {
+                amazonS3.deleteObject(bucket, formalS3Key);
+            }
         }
 
         // 저장
@@ -214,10 +216,12 @@ public class MypageService {
         metadata.setContentType("application/pdf");
 
         // 기존 파일 삭제
-        String formalS3Key = productRepo.findById(id).getDescriptionPath();
+        if(productRepo.findById(id).getDescriptionPath() != null) {
+            final String formalS3Key = productRepo.findById(id).getDescriptionPath();
 
-        if(amazonS3.doesObjectExist(bucket, formalS3Key)) {
-            amazonS3.deleteObject(bucket, formalS3Key);
+            if(amazonS3.doesObjectExist(bucket, formalS3Key)) {
+                amazonS3.deleteObject(bucket, formalS3Key);
+            }
         }
 
         // 저장
@@ -329,6 +333,17 @@ public class MypageService {
     }
 
     public void delete(final UserEntity userEntity) {
+        // 대본 삭제
+        for(ProductEntity product : productRepo.findAllByUserId(userEntity.getId())) {
+            if(amazonS3.doesObjectExist(bucket, product.getFilePath())) {
+                amazonS3.deleteObject(bucket, product.getFilePath());
+            }
+        }
+
+        // 작품 관련 이미지(scriptImage, description) 삭제
+
+
+        // DB 계정 삭제
         userRepo.delete(userEntity);
     }
 }
